pyproject.toml
[project]
name = "systtechbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0.0,<4.0.0",
    "openai>=1.0.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "psycopg[binary]>=3.1.0,<4.0.0",
    "psycopg-pool>=3.1.0,<4.0.0",
    "alembic>=1.13.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",           # coverage reporting
    "ruff>=0.1.0",                  # форматтер + линтер (замена black+flake8+isort)
    "mypy>=1.7.0",                  # type checking
]
api = [
    "fastapi>=0.104.0,<1.0.0",     # современный async веб-фреймворк
    "uvicorn[standard]>=0.24.0,<1.0.0",  # ASGI сервер
    "faker>=20.0.0,<21.0.0",       # генерация mock данных
    "httpx>=0.25.0",               # для тестирования FastAPI
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (ruff format handles this)

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

